Working database retrieval and update
