+++ src/pages/loginPage.js
@@ -1,4 +1,7 @@
 import React, { useState } from 'react';
+
+import '../firebaseConfig';
+
 import { getAuth, signInWithEmailAndPassword } from "firebase/auth";
 
 //Components
@@ -6,11 +9,11 @@ import PageTitle from '../components/PageTitle';
 import Description from '../components/Description';
 import Footer from '../components/Footer';
 
+const auth = getAuth();
+
 const LoginPage = (props) => {
     const [email, setEmail] = useState("");
     const [password, setPassword] = useState("");
-
-    const auth = getAuth();
 
     function prepareSignIn() {
         console.log("Starting signin");
