+++ src/pages/priorityPage.js
@@ -1,37 +1,41 @@
 import React, { useEffect, useState } from 'react';
 
+import '../firebaseConfig';
+
 import { getAuth, onAuthStateChanged, signOut } from "firebase/auth";
 import { getDatabase, ref, set, onValue } from "firebase/database";
 
 //Components
 import PageTitle from '../components/PageTitle';
-import Priorities from '../components/Priorities';
 import Description from '../components/Description';
 import Footer from '../components/Footer';
+import PriorityCard from "../components/PriorityCard";
+
+const auth = getAuth();
+const database = getDatabase();
 
 const PriorityPage = () => {
     const [priorityList, setPriorityList] = useState([]);
     const [priorityInput, setPriorityInput] = useState("");
     const [user, setUser] = useState(undefined);
-    const auth = getAuth();
-    const database = getDatabase();
 
-
-    onAuthStateChanged(auth, (userResult) => {
-        if (userResult) {
-            console.log("User found");
-            setUser(userResult);
-            onValue(ref(database, 'users/' + userResult.uid + '/priorities'), (snapshot) => {
-                const data = snapshot.val();
-                console.log("Loaded prio data: " + data);
-                setPriorityList(data);
-            });
-        } else {
-            console.log("User not found");
-            setUser(undefined);
-            window.location = "/login";
-        }
-    });
+    useEffect(() => {
+        onAuthStateChanged(auth, (userResult) => {
+            if (userResult) {
+                console.log("User found");
+                setUser(userResult);
+                onValue(ref(database, 'users/' + userResult.uid + '/priorities'), (snapshot) => {
+                    const data = snapshot.val();
+                    console.log("Loaded prio data: " + data);
+                    setPriorityList(data);
+                });
+            } else {
+                console.log("User not found");
+                setUser(undefined);
+                window.location = "/login";
+            }
+        });
+    }, [auth]);
 
     function writePriorities(list_of_priorities) {
         console.log("Writing! " + priorityList);
@@ -110,7 +114,9 @@ const PriorityPage = () => {
                     <button onClick={() => { userSignOut(); }} >Sign Out</button>
                     <input value={priorityInput} onChange={field => onPriorityInputChange(field.target.value)} type="text" id="priority_field" />
                     <button id="PriorityButton" onClick={() => addPriority()}>Add priority!</button>
-                    <Priorities priorityList={priorityList} key={priorityList} movePriority={movePriority} deletePriority={deletePriority} updatePriority={updatePriority} />
+                    {
+                        priorityList.map((priorityTitle, index) => <PriorityCard title={priorityTitle} key={index} priorityIndex={index} movePriority={movePriority} deletePriority={deletePriority} updatePriority={updatePriority} />)
+                    }
                 </div>
                 <Description />
                 <Footer />
